feat: 优惠券

+++ database/migrations/2024_09_01_092530_create_coupons_table.php
@@ -17,10 +17,10 @@ class CreateCouponsTable extends Migration
             $table->id();
             $table->integer('status')->comment('优惠券状态：1-有效，2-过期，3-下架');
             $table->string('name')->comment('优惠券名称');
-            $table->integer('denomination')->comment('优惠券面额');
+            $table->float('denomination')->comment('优惠券面额');
             $table->string('description')->comment('优惠券说明');
             $table->integer('num_limit')->default(0)->comment('优惠券商品数量门槛');
-            $table->integer('price_limit')->default(0)->comment('优惠券价格门槛');
+            $table->float('price_limit')->default(0)->comment('优惠券价格门槛');
             $table->string('expiration_time')->default('')->comment('优惠券失效时间');
             $table->integer('goods_id')->default(0)->comment('商品id');
             $table->string('goods_cover')->default('')->comment('商品图片');
